+++ src/main/host/syscall/handler/mod.rs
@@ -361,6 +361,7 @@ impl SyscallHandler {
             SyscallNum::NR_brk => handle!(brk),
             SyscallNum::NR_capget => handle!(capget),
             SyscallNum::NR_capset => handle!(capset),
+            SyscallNum::NR_chdir => handle!(chdir),
             SyscallNum::NR_clock_getres => handle!(clock_getres),
             SyscallNum::NR_clock_nanosleep => handle!(clock_nanosleep),
             SyscallNum::NR_clone => handle!(clone),