+++ src/test/unistd/test_unistd.rs
@@ -5,6 +5,8 @@
 
 use std::ffi::CStr;
 use std::ffi::CString;
+use std::ffi::OsStr;
+use std::os::unix::ffi::OsStrExt;
 use std::process;
 use std::sync::atomic::{AtomicUsize, Ordering};
 
@@ -53,6 +55,9 @@ fn main() {
     test_getpid_kill();
     test_getpgrp();
     test_getsid();
+    test_mkdir();
+    test_mkdirat();
+    test_chdir();
 }
 
 /// Tests that the results are plausible, but can't really validate that it's our
@@ -149,6 +154,177 @@ fn test_getpid_kill() {
     let rv = unsafe { libc::kill(pid as libc::pid_t, libc::SIGUSR1) };
     assert_eq!(rv, 0);
     assert_eq!(SIGACTION_COUNT.load(Ordering::SeqCst), 1);
+}
+
+fn test_mkdir() {
+    let dirname = c"./newdir";
+    let dirpath = OsStr::from_bytes(dirname.to_bytes());
+    let mode = libc::S_IRWXU;
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::mkdir(dirname.as_ptr(), mode)
+    })
+    .unwrap();
+
+    // Validate that the directory exists with the expected properties.
+    let mut stat: libc::stat = unsafe { std::mem::zeroed() };
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::stat(dirname.as_ptr(), &mut stat)
+    })
+    .unwrap();
+    // Validate that it has the expected permissions.
+    assert_eq!(stat.st_mode, libc::S_IFDIR | mode);
+
+    // Trying to create it again should fail.
+    assert_eq!(
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdir(dirname.as_ptr(), 0)
+        }),
+        Err(linux_api::errno::Errno::EEXIST)
+    );
+
+    // Clean up.
+    std::fs::remove_dir(dirpath).unwrap();
+
+    // Creating a directory with an invalid path component should fail.
+    assert_eq!(
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdir(c"./nonexistent/foo".as_ptr(), 0)
+        }),
+        Err(linux_api::errno::Errno::ENOENT)
+    );
+}
+
+fn test_mkdirat() {
+    // Create an outer directory first, using AT_FDCWD (equivalent to `mkdir`).
+    let outerdirname = c"./outerdir";
+    {
+        let mode = libc::S_IRWXU;
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdirat(libc::AT_FDCWD, outerdirname.as_ptr(), mode)
+        })
+        .unwrap();
+    }
+
+    let dirfd = linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::open(outerdirname.as_ptr(), libc::O_DIRECTORY)
+    })
+    .unwrap();
+    let basename = c"innerdir";
+    let mode = libc::S_IRWXU;
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::mkdirat(dirfd, basename.as_ptr(), mode)
+    })
+    .unwrap();
+
+    // Validate that the directory exists with the expected properties.
+    let mut stat: libc::stat = unsafe { std::mem::zeroed() };
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::stat(c"./outerdir/innerdir".as_ptr(), &mut stat)
+    })
+    .unwrap();
+    // Validate that it has the expected permissions.
+    assert_eq!(stat.st_mode, libc::S_IFDIR | mode);
+
+    // Trying to create it again should fail.
+    assert_eq!(
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdirat(dirfd, basename.as_ptr(), 0)
+        }),
+        Err(linux_api::errno::Errno::EEXIST)
+    );
+
+    // Clean up.
+    std::fs::remove_dir("./outerdir/innerdir").unwrap();
+    std::fs::remove_dir("./outerdir").unwrap();
+
+    // Creating a directory with an invalid path component should fail.
+    assert_eq!(
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdirat(dirfd, c"nonexistent/foo".as_ptr(), 0)
+        }),
+        Err(linux_api::errno::Errno::ENOENT)
+    );
+
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe { libc::close(dirfd) }).unwrap();
+
+    // Creating a directory with an invalid file descriptor should fail.
+    assert_eq!(
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdirat(-1, basename.as_ptr(), 0)
+        }),
+        Err(linux_api::errno::Errno::EBADF)
+    );
+
+    // Using a non-directory fd should fail.
+    let tmpfd = {
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::open(
+                c"regfile".as_ptr(),
+                libc::O_CREAT | libc::O_RDWR,
+                libc::S_IRWXU,
+            )
+        })
+        .unwrap()
+    };
+    assert_eq!(
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::mkdirat(tmpfd, basename.as_ptr(), 0)
+        }),
+        Err(linux_api::errno::Errno::ENOTDIR)
+    );
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe { libc::close(tmpfd) }).unwrap();
+    std::fs::remove_file("regfile").unwrap();
+}
+
+fn test_chdir() {
+    let start_cwd = std::env::current_dir().unwrap();
+
+    let mut new_dir = start_cwd.clone();
+    new_dir.push("newdir");
+
+    let new_dir_cstr = {
+        let mut v = Vec::from(new_dir.as_os_str().as_bytes());
+        v.push(0);
+        CString::from_vec_with_nul(v).unwrap()
+    };
+
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::mkdir(new_dir_cstr.as_ptr(), libc::S_IRWXU)
+    })
+    .unwrap();
+    linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+        libc::chdir(new_dir_cstr.as_ptr())
+    })
+    .unwrap();
+
+    // Check that cwd matches the new directory.
+    assert_eq!(&std::env::current_dir().unwrap(), &new_dir);
+
+    // Also check that a file created relative to cwd uses the new directory.
+    // This is specifically to look for problems related to
+    // <https://github.com/shadow/shadow/issues/2960>
+    {
+        // Create the file.
+        let path = c"./tmpfile";
+        let fd = linux_api::errno::Errno::result_from_libc_errno(-1, unsafe {
+            libc::open(path.as_ptr(), libc::O_CREAT | libc::O_RDWR, libc::S_IRWXU)
+        })
+        .unwrap();
+        // Close it.
+        linux_api::errno::Errno::result_from_libc_errno(-1, unsafe { libc::close(fd) }).unwrap();
+
+        // Validate that it exists at the expected path.
+        let mut expected_path = new_dir.clone();
+        expected_path.push("tmpfile");
+        std::fs::metadata(&expected_path).unwrap();
+
+        // Remove the file.
+        std::fs::remove_file(expected_path).unwrap();
+    }
+
+    // Clean up.
+    std::env::set_current_dir(start_cwd).unwrap();
+    std::fs::remove_dir(&new_dir).unwrap();
 }
 
 fn gethostname_with_short_buffer() -> libc::c_int {